requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   mcp
    #   openai
    #   sse-starlette
    #   starlette
asyncio==3.4.3
    # via fractflow (pyproject.toml)
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
distro==1.9.0
    # via openai
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.12
    # via fractflow (pyproject.toml)
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   fractflow (pyproject.toml)
    #   mcp
    #   openai
httpx-sse==0.4.0
    # via
    #   fractflow (pyproject.toml)
    #   mcp
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via fractflow (pyproject.toml)
jiter==0.9.0
    # via openai
markupsafe==3.0.2
    # via jinja2
mcp==1.5.0
    # via fractflow (pyproject.toml)
openai==1.68.2
    # via fractflow (pyproject.toml)
orjson==3.10.15
    # via fractflow (pyproject.toml)
pydantic==2.10.6
    # via
    #   fractflow (pyproject.toml)
    #   fastapi
    #   mcp
    #   openai
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.8.1
    # via mcp
python-dotenv==1.0.1
    # via
    #   fractflow (pyproject.toml)
    #   pydantic-settings
requests==2.32.3
    # via fractflow (pyproject.toml)
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sse-starlette==2.2.1
    # via mcp
starlette==0.46.1
    # via
    #   fastapi
    #   mcp
    #   sse-starlette
tenacity==9.0.0
    # via fractflow (pyproject.toml)
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   fractflow (pyproject.toml)
    #   anyio
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   uvicorn
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via
    #   fractflow (pyproject.toml)
    #   mcp
websockets==15.0.1
    # via fractflow (pyproject.toml)
loguru
pyyaml
json_repair
tokencost
opencv-python
flask
torch
segment_anything
torchvision
